+++ src/pages/admin/Appointments/Appointments.module.scss
@@ -1,97 +1,64 @@
+
+.title {
+  font-weight: 400;
+  font-size: 40px;
+  padding: 30px 0 100px 40px
+}
+
 .appointments {
-  padding: 20px;
-}
-.employers {
-  display: flex;
-  flex-direction: row;
-  justify-content: left;
-}
-.positionFlex {
-  display: flex;
-  align-items: center;
+  grid-template-columns: repeat(8, 1fr);
+  padding: 3%;
+  display: grid;
+  row-gap: 10px;
+  //gap: 20px;
+
 }
 
-.selectedDate {
-  margin-left: 35%;
-}
-
-.time {
-  height: 120px;
-  background-color: blue;
-  color: white;
-}
-
-.timeColor {
-  height: 120px;
-  background-color: red;
-  color: white;
-}
-
-.dateFlex {
-  display: flex;
-  justify-content: space-between;
-}
-
-.today {
-  border: 1px solid red;
-  cursor: pointer;
-  margin-left: 250px;
-}
-
-.row {
-  padding: 0 7%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
 .cell {
-  background: #FFFFFF;
-  box-shadow: 0px 0px 10px rgba(255, 153, 255, 0.2);
-  border-radius: 8px;
-}
-.button{
-  width: 36px;
-  height: 36px;
-
-  /* pink */
-
-  background: #D78DFF;
-  border-radius: 60px;
-  border: none;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.weekHeader {
-  display: flex;
-  width: 30%;
+  display: grid;
+  place-items: center;
+  padding: 10px 0 10px 10px;
+  font-size: 16px;
+  border: 1px solid #d89ff6;
   justify-content: space-between;
 
-  font-style: normal;
-  font-weight: 600;
-  font-size: 24px;
-  line-height: 130%;
-  color: #4F4E4E;
+  &:not(&:first-child) {
+    border-left: 1px solid #d89ff6;
+  }
 }
-.calendarRow {
+
+.startCell {
+  border-left: 1px solid #d89ff6;
+  border-radius: 40px 0 0 40px;
+
+}
+
+.endCell {
+  border-left: 1px solid #d89ff6;
+  border-radius: 0 40px 40px 0;
+}
+
+.rowTable {
+
+}
+.message {
+  padding: 20px;
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 10px;
-  gap: 10px;
-
-  width: 100%;
-  height: 42px;
-
-  background: #F9ECFF;
-  border-radius: 6px;
+  max-width: 350px;
+  min-height: 200px;
 }
-.buttonLeft {
-  transform: matrix(-1, 0, 0, -1, 0, 0);
+.openOption {
+  color: blue; /* Установите цвет для опции "Открыто" */
 }
 
-.buttonRight {
-  transform: matrix(1, 0, 0, -1, 0, 0);
+.inProgressOption {
+  color: green; /* Установите цвет для опции "В работе" */
+}
+
+.completedOption {
+  color: purple; /* Установите цвет для опции "Завершено" */
+}
+
+.canceledOption {
+  color: red; /* Установите цвет для опции "Отменено" */
 }
