+++ src/pages/admin/Appointments_old/Appointments.module.scss
@@ -1,0 +1,97 @@
+.appointments {
+  padding: 20px;
+}
+.employers {
+  display: flex;
+  flex-direction: row;
+  justify-content: left;
+}
+.positionFlex {
+  display: flex;
+  align-items: center;
+}
+
+.selectedDate {
+  margin-left: 35%;
+}
+
+.time {
+  height: 120px;
+  background-color: blue;
+  color: white;
+}
+
+.timeColor {
+  height: 120px;
+  background-color: red;
+  color: white;
+}
+
+.dateFlex {
+  display: flex;
+  justify-content: space-between;
+}
+
+.today {
+  border: 1px solid red;
+  cursor: pointer;
+  margin-left: 250px;
+}
+
+.row {
+  padding: 0 7%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+.cell {
+  background: #FFFFFF;
+  box-shadow: 0px 0px 10px rgba(255, 153, 255, 0.2);
+  border-radius: 8px;
+}
+.button{
+  width: 36px;
+  height: 36px;
+
+  /* pink */
+
+  background: #D78DFF;
+  border-radius: 60px;
+  border: none;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.weekHeader {
+  display: flex;
+  width: 30%;
+  justify-content: space-between;
+
+  font-style: normal;
+  font-weight: 600;
+  font-size: 24px;
+  line-height: 130%;
+  color: #4F4E4E;
+}
+.calendarRow {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 12px 10px;
+  gap: 10px;
+
+  width: 100%;
+  height: 42px;
+
+  background: #F9ECFF;
+  border-radius: 6px;
+}
+.buttonLeft {
+  transform: matrix(-1, 0, 0, -1, 0, 0);
+}
+
+.buttonRight {
+  transform: matrix(1, 0, 0, -1, 0, 0);
+}
